add testing.py module, refactor vasp tests and reduce code duplications
